LayoutTests/media/encrypted-media/encrypted-media-append-encrypted-unencrypted-expected.txt


EME API is supported OK
Appending chunks...
EVENT(encrypted)
MediaKeys is created OK
EVENT(message)
EVENT(keystatuseschange)
Session: keyId=b0b1b2b3b4b5b6b7b8b9babbbcbdbebf status=usable OK
EVENT(encrypted)
EVENT(message)
EVENT(keystatuseschange)
Session: keyId=b0b1b2b3b4b5b6b7b8b9babbbcbdbebf status=usable OK
All appends are completed
EXPECTED (video.duration >= '15') OK
END OF TEST
